refactor(interceptor): improve websocket utilities path
