+++ packages/zimic-interceptor/src/cli/server/__tests__/cli.serverStart.authentication.node.test.ts
@@ -11,7 +11,7 @@ import {
   listInterceptorTokens,
   removeInterceptorToken,
 } from '@/server/utils/auth';
-import UnauthorizedWebSocketConnectionError from '@/webSocket/errors/UnauthorizedWebSocketConnectionError';
+import UnauthorizedWebSocketConnectionError from '@/utils/webSocket/errors/UnauthorizedWebSocketConnectionError';
 import { usingIgnoredConsole } from '@tests/utils/console';
 import { usingHttpInterceptor } from '@tests/utils/interceptors';
