+++ packages/zimic-interceptor/src/cli/server/__tests__/cli.serverStart.node.test.ts
@@ -13,8 +13,8 @@ import { verifyUnhandledRequestMessage } from '@/http/interceptor/__tests__/shar
 import { createHttpInterceptor } from '@/http/interceptor/factory';
 import { DEFAULT_SERVER_LIFE_CYCLE_TIMEOUT } from '@/server/constants';
 import { importCrypto } from '@/utils/crypto';
-import WebSocketClient from '@/webSocket/WebSocketClient';
-import WebSocketServer from '@/webSocket/WebSocketServer';
+import WebSocketClient from '@/utils/webSocket/WebSocketClient';
+import WebSocketServer from '@/utils/webSocket/WebSocketServer';
 import { usingIgnoredConsole } from '@tests/utils/console';
 import { usingHttpInterceptor } from '@tests/utils/interceptors';
