+++ packages/zimic-interceptor/src/utils/webSocket/errors/InvalidWebSocketMessageError.ts
