+++ packages/zimic-interceptor/src/utils/webSocket/errors/NotRunningWebSocketHandlerError.ts
