+++ packages/zimic-interceptor/src/utils/webSocket/errors/UnauthorizedWebSocketConnectionError.ts
