+++ packages/zimic-interceptor/src/utils/webSocket/types.ts
